+++ Magic/CHANGELOG.md
@@ -21,6 +21,7 @@
  - Fix using variables in the ModifyAttributes and ModifyProperties actions
  - Fix EffectLib's ExplodeEffect and BigBangEffect
  - Fix being able to put `unstashable` wands in chests if they were in your main hand
+ - Fix an error that could occor if a spell kills its own caster inside of an Asyncronous action
 
 ## Changes